Added Input Handlers
Changed config.ini file to also have the 12 bands set with YES/NO; Added methods for Input/Output so that the user can opt for YES/NO bands wanted for future download/process of the images; deleted information_gather.py because it was not necessary; At this point the full name of the files .tif, based on the specifications from the configuration files, is extracted. Need testing!