OK, it works! It calculates. Fully integrated the argparser, dumped configurator. Need refactor
Solved the seg fault bug. The bands are refferences to the tiffs. When computing in the same local method, the tiffs were not available globally anymore, so the bands had a refference to nothing. It would seg fault at numpy.ReadAsArray(...). Fully integrated argument parser. Just need some ( a lot) of refactoring.