Updated argument parser AAAAND (yes..) refactored a little bit of download.
Updates the argument parser so that it takes only the needed attributes. It has as bug when called from pycharm, AttributeError. Maybe because it's not instantiated as supposed; works fine from the command line. TODO solve the bug. Deleted start and end from the download file as it wasn't needed since the csv encode is solved, as well as threading.