Separated the NDSI caller form the argument file.
Created a class NDSI Caller which uses the argument line process parameters to setup the searching for the green and swir1 bands in the input directory, calclate the ndsi for the paired bands (if all bands are paired. Will fix this so if there are 99 pairs, and 1 is not, it will still calculate for that one.). Set up the threshold.