Improved multi-threading. BUG.
Improved the multithreading to handle ctrl+C as it should. Bug: when KeyInterrupt is called while the process is in process_alignment.py, it gives cannot import cv2 error, nump.core.multiarray failed to import. HOWTOHUGGINSOLVEIT: Find the method which does not closes to sig kill, and close it BY FOOOOOORCE.