Implemented align writing to csv. Started ndsi writing to csv. Resolved big big BIG BUGS.
Solved zero value / corrupted image error. When calculating keypoints and descriptors on corrupted images, the processing would stop raising an error.  Solved processing exiting before the last process is done, because the wait for finish was handled only on exception rather than actually finishing the job. Solved aligned csv writing to bad path and not writing correctly the first rows with the name of the attributes. Implemented csv in a differenc class and did some code refactoring for simplification.